+++ packages/functions/src/cache/index.ts
@@ -35,6 +35,9 @@ export const getCache = (cacheOptions?: CacheOptions): RuntimeCache => {
     if (ctxCache) return ctxCache as RuntimeCache;
     if (!inMemoryCacheInstance) {
       inMemoryCacheInstance = new InMemoryCache();
+      console.warn(
+        'Runtime Cache unavailable in this environment. Falling back to in-memory cache.'
+      );
     }
     return inMemoryCacheInstance;
   };